+++ libs/ros2bridge/src/time.cpp
@@ -12,14 +12,12 @@
 
 #include <cmath>  // std::fmod
 
-namespace mrpt::ros2bridge
-{
-mrpt::system::TTimeStamp fromROS(const rclcpp::Time& src)
+mrpt::system::TTimeStamp mrpt::ros2bridge::fromROS(const rclcpp::Time& src)
 {
 	return mrpt::Clock::fromDouble(src.seconds());
 }
 
-rclcpp::Time toROS(const mrpt::system::TTimeStamp& src)
+rclcpp::Time mrpt::ros2bridge::toROS(const mrpt::system::TTimeStamp& src)
 {
 	// Convert to "double-version of time_t", then extract integer and
 	// fractional parts:
@@ -30,4 +28,3 @@ rclcpp::Time toROS(const mrpt::system::TTimeStamp& src)
 		// nanoseconds:
 		static_cast<uint64_t>(std::fmod(t, 1.0) * 1e9 + 0.5 /*round*/));
 }
-}  // namespace mrpt::ros2bridge